Add hover tooltip for new-chat explore percent control
